refactor(decompose): split webapp/e2e-tests/electron/for_feature_development_not_LT_verification/screenshots/electron-slider-screenshots.spec.ts into folder
